Removed NUnit_TestRunner_Tests
As discussed with @FraserGreenroyd -- this test project was meant to replicate the conditions of BHoMBot invoking test methods. However, this test project shows repeated issues in loading required assemblies. After spending some time on it, we decided to drop it and verify whether test projects are run correctly directly via BHoMBot. Tests remain runnable by users via the Test Explorer, and the base NUnitTest class retains the feature of loading of required assemblies via the OnTimeSetup method; the OnTimeSetup is only invoked when the Build Config is set to "Test", which BHoMBot never uses.